refactor: move tracking to shared/ to avoid new circular imports
Move limitedSupportTracking.ts from model-serving/src/tracking/ to
model-serving/src/shared/tracking/ so it exports via the ./shared
path. This avoids introducing new restricted cross-package imports
from frontend/ into model-serving internals.
